android/requirements.txt
Appium-Python-Client==2.9.0
webdriver-manager==3.8.5
pytest==7.2.2
selenium==4.8.3